Correct Restore on a New Install
Restoring an archive and requesting existing items be deleted first could crash on a new install because it would attempt to use a bills folder that did not exist yet.

Updated `PermanentlyDeleteAllLocalMeals` to check if the meal folder exists before attempting to enumerate and delete files, preventing errors during new installations or when the folder is missing.

Deleted an unnecessary comment in Archive.

+++ DivisiBill/Models/MealSummary.cs
@@ -318,17 +318,20 @@ public partial class MealSummary : ObservableObjectPlus, IComparable<MealSummary
     /// an archive restore operation. All in-memory sample meals are also cleared.</remarks>
     public static void PermanentlyDeleteAllLocalMeals()
     {
-        var mealFiles = Directory.EnumerateFiles(Meal.MealFolderPath, "*.xml").ToList();
-        
-        foreach (string mealFilePath in mealFiles)
+        if (Directory.Exists(Meal.MealFolderPath)) // It may not exist if this is a new install
         {
-            try
+            var mealFiles = Directory.EnumerateFiles(Meal.MealFolderPath, "*.xml").ToList();
+
+            foreach (string mealFilePath in mealFiles)
             {
-                File.Delete(mealFilePath);
-            }
-            catch (Exception ex)
-            {
-                ex.ReportCrash();
+                try
+                {
+                    File.Delete(mealFilePath);
+                }
+                catch (Exception ex)
+                {
+                    ex.ReportCrash();
+                }
             }
         }
         Meal.LocalMealList.Clear(); // Will also delete any in-memory only sample meals
